requirements_streamlit.txt
streamlit>=1.28.0
pandas>=1.5.0
openpyxl>=3.0.0
python-pptx>=0.6.21
Pillow>=9.0.0 